Converted Global to Module (according to play upgrading guideline) and configured H2 datasource in spring configuration
