+++ app/modules/SpringDataJpaModule.java
@@ -1,0 +1,75 @@
+package modules;
+
+import org.springframework.context.annotation.AnnotationConfigApplicationContext;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
+import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
+import org.springframework.orm.hibernate3.HibernateExceptionTranslator;
+import org.springframework.orm.jpa.JpaTransactionManager;
+import org.springframework.guice.module.SpringModule;
+import org.springframework.orm.jpa.LocalContainerEntityManagerFactoryBean;
+import org.springframework.orm.jpa.vendor.HibernateJpaVendorAdapter;
+
+import javax.inject.Inject;
+import javax.inject.Singleton;
+import javax.persistence.EntityManagerFactory;
+import javax.persistence.Persistence;
+import javax.sql.DataSource;
+
+/**
+ * Application wide behaviour. We establish a Spring application context for the dependency injection system and
+ * configure Spring Data.
+ */
+
+@Singleton
+public class SpringDataJpaModule extends SpringModule {
+
+    /**
+     * The name of the persistence unit we will be using.
+     */
+    static final String DEFAULT_PERSISTENCE_UNIT = "default";
+
+    /**
+     * Sync the context lifecycle with Play's.
+     */
+    @Inject
+    public SpringDataJpaModule() {
+        super(new AnnotationConfigApplicationContext(SpringDataJpaConfiguration.class));
+    }
+    /**
+     * This configuration establishes Spring Data concerns including those of JPA.
+     */
+    @Configuration
+    @EnableJpaRepositories("models")
+    public static class SpringDataJpaConfiguration {
+
+        @Bean
+        public DataSource dataSource() {
+            return new EmbeddedDatabaseBuilder()
+                .setType(EmbeddedDatabaseType.H2)
+                .setName("DefaultDS")
+                .build();
+        }
+
+        @Bean
+        public LocalContainerEntityManagerFactoryBean entityManagerFactory() {
+            LocalContainerEntityManagerFactoryBean factory = new LocalContainerEntityManagerFactoryBean();
+            factory.setDataSource(dataSource());
+            factory.setPersistenceUnitName(DEFAULT_PERSISTENCE_UNIT);
+            factory.setJpaVendorAdapter(new HibernateJpaVendorAdapter());
+            return factory;
+        }
+
+        @Bean
+        public HibernateExceptionTranslator hibernateExceptionTranslator() {
+            return new HibernateExceptionTranslator();
+        }
+
+        @Bean
+        public JpaTransactionManager transactionManager() {
+            return new JpaTransactionManager();
+        }
+    }
+}
