+++ conf/application.conf
@@ -30,9 +30,7 @@ application.langs="en"
 
 # Database configuration
 #
-db.default.driver=org.h2.Driver
-db.default.url="jdbc:h2:mem:play"
-db.default.jndiName=DefaultDS
+play.modules.enabled+="modules.SpringDataJpaModule"
 
 # Evolutions
 # ~~~~~
